Class Image added
